rename as popoverHint, add keyboard shortcut for toggling it
